Fix clang-tidy Part 3

+++ projects/rocprofiler-systems/source/lib/core/trace_cache/perfetto_processor.cpp
@@ -621,8 +621,9 @@ perfetto_processor_t::handle(const scratch_memory_sample& _sms)
     auto        _corr_id         = _sms.correlation_id_internal;
     auto        _stream_id       = _sms.stream_handle;
     auto        _queue_id_handle = _sms.queue_id_handle;
-    const auto& _t_info          = thread_info::get(_sms.thread_id, SystemTID);
-    const auto  _thread_id_sequent =
+    const auto& _t_info =
+        thread_info::get(static_cast<std::int64_t>(_sms.thread_id), SystemTID);
+    const auto _thread_id_sequent =
         (_t_info && _t_info->index_data) ? _t_info->index_data->sequent_value : 0U;
     auto _beg_ts = _sms.start_timestamp;
     auto _end_ts = _sms.end_timestamp;
@@ -1450,6 +1451,7 @@ perfetto_processor_t::handle([[maybe_unused]] const ainic_pmc_sample& _nic_sampl
     }
 }
 
+// NOLINTBEGIN(readability-function-size)
 void
 perfetto_processor_t::handle([[maybe_unused]] const hipfile_pmc_sample& _hipfile_sample)
 {
@@ -1480,10 +1482,11 @@ perfetto_processor_t::handle([[maybe_unused]] const hipfile_pmc_sample& _hipfile
         }
 
         TRACE_COUNTER(trait::name<category::hipfile>::value,
-                      hipfile_track::at(_track_key, 0), static_cast<std::int64_t>(_ts),
-                      _metric.value(_hipfile_sample.metric_values));
+                      hipfile_track::at(_track_key, 0), _ts,
+                      static_cast<double>(_metric.value(_hipfile_sample.metric_values)));
     }
 }
+// NOLINTEND(readability-function-size)
 
 void
 perfetto_processor_t::handle([[maybe_unused]] const in_time_sample& _sample)
